Big Chances; small changes.
EndpointStats.sh and RabbitStats.sh have had big changes to help it behave better while being used in cron. I removed the echos for this reason as cron will try to send email whenever there is standard output.

MoveToNewNodeZero.sh has had extra checks applied to it to deal with a mongo that may be locked away with a different password. All the yums have been removed as it seems they are not required. In the KB that this will be attached to, I have defined that.