Create FindUglyRules.py
This is a working script (that needs exception handling) that I needed to come up with that will allow me to read through an ESA deployments custom rules and identify potential ones that may be bad on performance.  This script connects to the Mongo Database on the Admin Server and reads 2 collections and then parses the raw text of the rules to identify possible problems with performance.

It's not amazing but it's better than having to pivot through many rules in the UI to figure out what may be problematic.