Completed salt-group script. Updated UglyRules
The salt-group script that I left orphan for a very long time is finally finished. This script can be used to help Admins run tasks simultaneously via the command line.

I also updated FindUglyRules.py to help with some character encoding that can cause it to die. Apparently this is better practice and I should get in the habit of doing it more.