Create SpamUEBA.py
I am prototyping a script that will spam Windows event as configured by parameter in an attempt to setoff UEBA. In theory, many of these events would have to be sent and perhaps in different combinations to illicit certain responses. The idea would be to cron this script or loop it to generate the desired events.

UEBA is currently rebuilding in my environment so I can't verify if this will work or not as of yet.