Added 2 scripts
Today, I wrote FindDeviatingSessionIDs.sh to help locate when the 11.3 app rule bug causes the sessiondb to diverge when saving meta. With the information generated by this script, support engineers will know where to perform the metadb rollover.

I also threw in RebuildSARabbit11.sh just to complete the collection since my last day is November 27th. This script is meant to rebuild a SA head when the RabbitMQ configuration is either reset or corrupted. This should only be a last resort for people.